requirements.txt
python-dotenv
awscli
boto3
pytest
pyspark
logdna
bugsnag
